+++ src/app/(user)/relation/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Relation } from '@/components/relation/Relation'
+import { Relation } from '@/components/relation/Relation/Relation'
 
 export default function relationPage() {
   return (
