+++ src/components/relation/utils/domain.ts
